requirements.txt
flask
flask-cors
chromadb
pandas
numpy
sentence-transformers
google-generativeai
python-dotenv
werkzeug
gunicorn
